LayoutTests/imported/w3c/web-platform-tests/digital-credentials/historical.any.worker-expected.txt


PASS navigator.identity is obsolete
